+++ src/extension.ts
@@ -110,7 +110,6 @@ export function activate(context: vscode.ExtensionContext): void {
 export function deactive(): void {
     rmAllDiag();
     statusBarItem.hide(); // just .hide() not .dispose()
-    log.clear();
 }
 
 /*